Move to a private selector, use the metabox initialization hook instead
