+++ packages/core-data/src/selectors.ts
@@ -1600,18 +1600,6 @@ export const getRevision = createSelector(
 );
 
 /**
- * Returns true if collaboration was disabled at runtime (e.g. due to
- * incompatible meta boxes).
- *
- * @param state Data state.
- *
- * @return Whether collaboration has been disabled.
- */
-export function isCollaborationDisabled( state: State ): boolean {
-	return state.collaborationDisabled;
-}
-
-/**
  * Returns the current sync connection status across all entities. Prioritizes
  * disconnected states, then connecting, then connected.
  *
