+++ packages/edit-post/src/components/layout/index.js
@@ -24,7 +24,6 @@ import { PluginArea } from '@wordpress/plugins';
 import { __, sprintf } from '@wordpress/i18n';
 import {
 	useCallback,
-	useEffect,
 	useMemo,
 	useId,
 	useRef,
@@ -397,13 +396,10 @@ function Layout( {
 		isWelcomeGuideVisible,
 		templateId,
 		isDevicePreview,
-		areMetaBoxesReady,
-		isCollaborationEnabled,
 	} = useSelect(
 		( select ) => {
 			const { get } = select( preferencesStore );
-			const { isFeatureActive, hasMetaBoxes, areMetaBoxesInitialized } =
-				select( editPostStore );
+			const { isFeatureActive, hasMetaBoxes } = select( editPostStore );
 			const { canUser, getPostType, getTemplateId } = unlock(
 				select( coreStore )
 			);
@@ -451,11 +447,6 @@ function Layout( {
 						? _templateId
 						: null,
 				isDevicePreview: getDeviceType() !== 'Desktop',
-				areMetaBoxesReady: areMetaBoxesInitialized() && hasMetaBoxes(),
-				isCollaborationEnabled:
-					select(
-						editorStore
-					).isCollaborationEnabledForCurrentPost(),
 			};
 		},
 		[
@@ -468,21 +459,6 @@ function Layout( {
 	);
 
 	useMetaBoxInitialization( hasActiveMetaboxes && hasResolvedMode );
-
-	// Disable real-time collaboration when non-compatible meta boxes are detected.
-	const { disableCollaboration } = unlock( useDispatch( coreStore ) );
-
-	useEffect( () => {
-		if ( areMetaBoxesReady && isCollaborationEnabled ) {
-			disableCollaboration();
-		}
-	}, [
-		areMetaBoxesReady,
-		isCollaborationEnabled,
-		disableCollaboration,
-		currentPostType,
-		currentPostId,
-	] );
 
 	// Set the right context for the command palette
 	const commandContext = hasBlockSelected
